+++ .rubocop_todo.yml
@@ -701,7 +701,7 @@ RSpec/ContextWording:
     - 'spec/models/claim/base_claim_spec.rb'
     - 'spec/models/claim/litigator_claim_spec.rb'
     - 'spec/models/claim/litigator_hardship_claim_spec.rb'
-    - 'spec/models/claim/transfer_brain_data_item_collection_spec.rb'
+    - 'spec/models/claim/transfer_brain/data_item_collection_spec.rb'
     - 'spec/models/claim/transfer_claim_spec.rb'
     - 'spec/models/claim/transfer_detail_spec.rb'
     - 'spec/models/claim_state_transition_reason_spec.rb'
@@ -1125,7 +1125,7 @@ RSpec/ImplicitSubject:
     - 'spec/api/v2/ccr_claim_spec.rb'
     - 'spec/lib/gov_uk/date_accessor_spec.rb'
     - 'spec/models/claim/base_claim_spec.rb'
-    - 'spec/models/claim/transfer_brain_data_item_collection_spec.rb'
+    - 'spec/models/claim/transfer_brain/data_item_collection_spec.rb'
     - 'spec/models/claims/totals_spec.rb'
     - 'spec/models/fee/fixed_fee_type_spec.rb'
     - 'spec/models/fee/misc_fee_type_spec.rb'
@@ -1451,7 +1451,7 @@ RSpec/MultipleExpectations:
     - 'spec/models/claim/base_claim_spec.rb'
     - 'spec/models/claim/litigator_claim_spec.rb'
     - 'spec/models/claim/litigator_hardship_claim_spec.rb'
-    - 'spec/models/claim/transfer_brain_data_item_collection_spec.rb'
+    - 'spec/models/claim/transfer_brain/data_item_collection_spec.rb'
     - 'spec/models/claim/transfer_brain_spec.rb'
     - 'spec/models/claim/transfer_claim_spec.rb'
     - 'spec/models/claim_state_transition_reason_spec.rb'
@@ -1768,7 +1768,7 @@ RSpec/NotToNot:
     - 'spec/mailers/notify_mailer_spec.rb'
     - 'spec/models/allocation_spec.rb'
     - 'spec/models/claim/base_claim_spec.rb'
-    - 'spec/models/claim/transfer_brain_data_item_collection_spec.rb'
+    - 'spec/models/claim/transfer_brain/data_item_collection_spec.rb'
     - 'spec/models/claim/transfer_detail_spec.rb'
     - 'spec/models/claims/cloner_spec.rb'
     - 'spec/models/claims/financial_summary_spec.rb'
@@ -1857,7 +1857,7 @@ RSpec/RepeatedExampleGroupDescription:
   Exclude:
     - 'spec/models/ability_spec.rb'
     - 'spec/models/claim/base_claim_spec.rb'
-    - 'spec/models/claim/transfer_brain_data_item_collection_spec.rb'
+    - 'spec/models/claim/transfer_brain/data_item_collection_spec.rb'
     - 'spec/models/fee/interim_fee_spec.rb'
     - 'spec/presenters/management_information_presenter_spec.rb'
     - 'spec/services/claims/fee_calculator/graduated_price_spec.rb'
@@ -3210,7 +3210,7 @@ Layout/LineLength:
     - 'spec/models/claim/advocate_claim_spec.rb'
     - 'spec/models/claim/base_claim_spec.rb'
     - 'spec/models/claim/litigator_claim_spec.rb'
-    - 'spec/models/claim/transfer_brain_data_item_collection_spec.rb'
+    - 'spec/models/claim/transfer_brain/data_item_collection_spec.rb'
     - 'spec/models/claim/transfer_brain_spec.rb'
     - 'spec/models/claim/transfer_claim_spec.rb'
     - 'spec/models/claim/transfer_detail_spec.rb'
