+++ app/models/concerns/transfer_brain_data_item_collection_delegatable.rb
@@ -5,7 +5,7 @@ module TransferBrainDataItemCollectionDelegatable
     def data_item_collection_delegate(*method_names)
       method_names.each do |method_name|
         define_singleton_method(method_name) do |detail|
-          Claim::TransferBrainDataItemCollection.instance.send(method_name.to_sym, detail)
+          Claim::TransferBrain::DataItemCollection.instance.send(method_name.to_sym, detail)
         end
       end
     end
